% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AD_frequency}
\alias{AD_frequency}
\title{AD_frequency}
\arguments{
\item{ad}{a matrix of allele depths (e.g., "7,2")}

\item{allele}{which (1-based) allele to report frequency for}

\item{sum_type}{type of sum to calculate, see details}

\item{delim}{character that delimits values}
}
\value{
a numeric matrix of frequencies
[[Rcpp::export]]
}
\description{
Create allele frequencies from matrices of allelic depths (AD)
}
\details{
To do.

There are two methods to calculate a sum for the denominator of the frequency.
When sum_type = 0 the alleles are sorted decendingly and the first two allele counts are used for the sum.
This may be useful when a state of diploidy may be known to be appropriate and other alleles may be interpreted as erroneous.
When sum_type = 1 a sum is taken over all the observed alleles for a variant.
}

