% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addID.R
\name{addID}
\alias{addID}
\title{Populate the ID column of VCF data}
\usage{
addID(x)
}
\arguments{
\item{x}{an object of class vcfR or chromR}
}
\description{
Populate the ID column of VCF data by concatenating the chromosome, position and optionally an index.
}
\details{
Variant callers typically leave the ID column empty in VCF data.
However, the VCF data may potentially include variants with IDs as well as variants without.
This function populates the missing elements by concatenating the chromosome and position.
When this concatenation results in non-unique names, an index is added to force uniqueness.
}

