\docType{class}
\name{Chrom-class}
\alias{Chrom-class}
\title{Chrom class}
\description{
A class for representing chromosomes (or contigs).
}
\details{
Defines a class for chromosomal or contig data.

for variant call format data. A vcfR object contains three
slots.  The first slot is a character vector which holds
the meta data.  The second slot holds an eight column
data.frame to hold the fixed data.  The third slot is a
data.frame which holds the genotype data.
}

