% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Chrom_methods.R
\docType{methods}
\name{show,Chrom-method}
\alias{head,Chrom-method}
\alias{head.chrom}
\alias{names,Chrom-method}
\alias{names.chrom}
\alias{plot,Chrom-method}
\alias{plot.chrom}
\alias{show,Chrom-method}
\alias{windowize}
\title{Chrom methods}
\usage{
\S4method{show}{Chrom}(object)

\S4method{names}{Chrom}(x)

\S4method{head}{Chrom}(x)

\S4method{plot}{Chrom}(x, y, ...)

windowize(x, win.size = 1000, max.win = 10000)
}
\arguments{
\item{object}{an object of class Chrom}

\item{x}{an object of class Chrom}

\item{y}{some sort of object???}

\item{...}{Arguments to be passed to methods}

\item{win.size}{window size, in base pairs}

\item{max.win}{maximum window size}
}
\description{
Chrom methods
}
\details{
Reads in a vcf file and stores it in a Chrom class.
}

