% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chromo_plot.R
\name{chromo_plot}
\alias{chromo}
\alias{chromo_plot}
\alias{chromodot}
\alias{chromohwe}
\alias{chromopop}
\alias{chromoqc}
\title{Plot chromR object}
\usage{
chromo(chrom, ...)

chromoqc(chrom, boxp = TRUE, alpha = 255, ...)

chromohwe(chrom, ...)

chromodot(chrom, ...)

chromopop(chrom, ...)
}
\arguments{
\item{chrom}{an object of class chrom.}

\item{...}{arguments to be passed to other methods.}

\item{boxp}{logical specifying whether marginal boxplots should be plotted.}

\item{alpha}{degree of transparency applied to points in dot plots.}
}
\description{
Plot chromR object
}

