% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chromo_plot2.R
\name{chromo plot v2}
\alias{bar.plot}
\alias{chromo plot v2}
\alias{dot.plot}
\alias{rect.plot}
\title{chromo plot v2}
\usage{
dot.plot(mat, title = NULL, hline = NULL, col = NULL, hist = TRUE,
  layout = TRUE, mwidth = 4, ...)

bar.plot(mat, scale = FALSE, title = NULL, col = NULL, hist = TRUE,
  layout = TRUE, mwidth = 4, ...)

rect.plot(mat, ...)
}
\arguments{
\item{mat}{a numeric matrix where column one represents the chromosomal position (POS) and subsequent columns represent y-values}

\item{title}{a title for the plot}

\item{hline}{numeric vector specifying positions for horizontal lines to be drawn}

\item{col}{a vector of colors for each column of y-values, recycled as necessary}

\item{hist}{logical specifying whether marginal histograms should be plotted}

\item{layout}{logical specifying whether to call layout or not}

\item{mwidth}{numeric specifying the relative width of the main panel when using layout}

\item{...}{arguments to be passed to/from other methods}

\item{scale}{logical specifying whether to scale bars to one}
}
\description{
Plot chromR objects and their components
}
\details{


Plot chromR objects and their components.

The function \strong{dot.plot}

The function \strong{bar.plot}

The function \strong{rect.plot}
}

