% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromR-method.R
\docType{methods}
\name{show,chromR-method}
\alias{chromR-method}
\alias{getName}
\alias{getName,chromR-method}
\alias{head,chromR-method}
\alias{head.chromR}
\alias{plot,chromR-method}
\alias{plot.chromR}
\alias{setName}
\alias{setName<-,chromR-method}
\alias{show,chromR-method}
\alias{windowize}
\title{chromR-method}
\usage{
\S4method{show}{chromR}(object)

\S4method{head}{chromR}(x)

\S4method{plot}{chromR}(x, y, ...)

\S4method{getName}{chromR}(object)

\S4method{setName}{chromR}(object) <- value

windowize(x, win.size = 1000, max.win = 10000)
}
\arguments{
\item{object}{an object of class chromR}

\item{x}{an object of class chromR}

\item{y}{some sort of object???}

\item{...}{Arguments to be passed to methods}

\item{win.size}{window size, in base pairs}

\item{max.win}{maximum window size}
}
\description{
Methods that act on objects of class chromR
}
\details{
Methods that act on objects of class chromR.
}
\examples{

library(vcfR)
data(vcfR_example)
chrom <- create.chromR('sc50', seq=dna, vcf=vcf, ann=gff)
chrom

print(chrom)
head(chrom)
plot(chrom)
setName(chrom) <- "Supercontig_1.50"
getName(chrom)
names(chrom) <- "Scaffold_50"
names(chrom)


}

