% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromR_example.R
\docType{data}
\name{chromR_example}
\alias{chromR_example}
\alias{chrom}
\title{Example chromR object.}
\format{A chromR object}
\description{
An example chromR object containing parts of the *Phytophthora infestans* genome.
}
\details{
This data is a subset of the pinfsc50 dataset.
It has been subset to positions between 500 and 600 kbp.
The coordinate systems of the vcf and gff file have been altered by subtracting 500,000.
This results in a 100 kbp section of supercontig_1.50 that has positional data ranging from 1 to 100 kbp.
}
\examples{
data(chromR_example)




}
\keyword{datasets}
