% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chromR_methods.R
\docType{methods}
\name{show,chromR-method}
\alias{head,chromR-method}
\alias{head.chromR}
\alias{names,chromR-method}
\alias{names.chromR}
\alias{plot,chromR-method}
\alias{plot.chromR}
\alias{show,chromR-method}
\alias{windowize}
\title{chromR methods}
\usage{
\S4method{show}{chromR}(object)

\S4method{names}{chromR}(x)

\S4method{head}{chromR}(x)

\S4method{plot}{chromR}(x, y, ...)

windowize(x, win.size = 1000, max.win = 10000)
}
\arguments{
\item{object}{an object of class chromR}

\item{x}{an object of class chromR}

\item{y}{some sort of object???}

\item{...}{Arguments to be passed to methods}

\item{win.size}{window size, in base pairs}

\item{max.win}{maximum window size}
}
\description{
chromR methods
}
\details{
Reads in a vcf file and stores it in a chromR class.
}

