% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_chromR.R
\name{Create chromR object}
\alias{Create chromR object}
\alias{ann2chromR}
\alias{chromR-methods}
\alias{create_chromR}
\alias{getCHROM}
\alias{getDP}
\alias{getFIX}
\alias{getINFO}
\alias{getPOS}
\alias{getQUAL}
\alias{seq2chromR}
\alias{vcf2chromR}
\title{Create chromR object}
\usage{
create_chromR(name = "CHROM1", vcf, seq = NULL, ann = NULL,
  verbose = TRUE)

vcf2chromR(x, vcf)

seq2chromR(x, seq = NULL)

ann2chromR(x, gff)

getFIX(x)

getCHROM(x)

getPOS(x)

getQUAL(x)

getDP(x)

getINFO(x, element = "MQ")
}
\arguments{
\item{name}{a name for the chromosome (for plotting purposes)}

\item{vcf}{an object of class vcfR}

\item{seq}{a sequence as a DNAbin object}

\item{ann}{an annotation file (gff-like)}

\item{verbose}{should verbose output be printed to the console?}

\item{x}{an object of class chromR}

\item{gff}{a data.frame containing annotation data in the gff format}

\item{element}{element to extract from chromR object}
}
\description{
Creates and populates an object of class chromR.
}
\details{
Creates and names a chromR object from a name, a chromosome (an ape::DNAbin object), variant data (a vcfR object) and annotation data (gff-like).
The function \strong{create_chromR} is a wrapper which calls functions to populate the slots of the chromR object.

The function \strong{vcf2chromR} is called by create_chromR and transfers the data from the slots of a vcfR object to the slots of a chromR object.
It also tries to extract the 'DP' and 'MQ' fileds (when present) from teh fix region's INFO column.
It is not anticipated that a user would need to use this function directly, but its placed here in case they do.

The function \strong{seq2chromR} is currently defined as a generic function.
This may change in the future.
This function takes an object of class DNAbin and assigns it to the 'seq' slot of a chromR object.

The function \strong{ann2chromR} is called by create_chromR and transfers the information from a gff-like object to the 'ann' slot of a chromR object.
It is not anticipated that a user would need to use this function directly, but its placed here in case they do.
}
\examples{
library(vcfR)
}
\seealso{
\code{\link{chromR-class}},
\code{\link{vcfR-class}},
\code{\link[ape]{DNAbin}},
\href{http://www.1000genomes.org/wiki/analysis/variant\%20call\%20format/vcf-variant-call-format-version-41}{vcf format},
\href{http://www.sequenceontology.org/gff3.shtml}{gff3 format}
}

