% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drplot.R
\name{dr.plot elements}
\alias{dr.plot}
\alias{dr.plot elements}
\alias{null.plot}
\title{dr.plot elements}
\usage{
dr.plot(dmat = NULL, rlst = NULL, chrom.s = 1, chrom.e = NULL,
  title = NULL, hline = NULL, dcol = NULL, rcol = NULL, rbcol = NULL,
  ...)

null.plot()
}
\arguments{
\item{dmat}{a numeric matrix for dot plots where the first column is position (POS) and subsequent columns are y-values.}

\item{rlst}{a list containing numeric matrices containing rectangle coordinates.}

\item{chrom.s}{start position for the chromosome}

\item{chrom.e}{end position for the chromosome}

\item{title}{optional string to be used for the plot title.}

\item{hline}{vector of positions to be used for horizontal lines.}

\item{dcol}{vector of colors to be used for dot plots.}

\item{rcol}{vector of colors to be used for rectangle plots.}

\item{rbcol}{vector of colors to be used for rectangle borders.}

\item{...}{arguments to be passed to other methods.}
}
\value{
Returns the y-axis minimum and maximum values invisibly.
}
\description{
Plot chromR objects and their components
}
\details{
Plot details
The parameter \strong{rlist} is list of numeric matrices containing rectangle coordinates.
The first column of each matrix is the left positions, the second column is the bottom coordinates, the third column is the right coordinates and the fourth column is the top coordinates.
}
\seealso{
\code{\link[graphics]{rect}}
\code{\link{chromo}}
}

