\docType{data}
\name{example_data}
\alias{example_data}
\alias{pinf_dna}
\alias{pinf_gff}
\alias{pinf_vcf}
\title{Example data for vcfR.}
\format{A DNAbin object, a data.frame and a vcfR object}
\description{
An example dataset containing parts of the Phytophthora
infestans genome.
}
\details{
\itemize{ \item pinf_dna mitochondion IIa (GenBank:
AY898627.1) as a DNAbin object (ape) \item pinf_gff
annotation data (gff-like) as a data.frame \item pinf_vcf
variant information as a vcfR object

}
}
\keyword{datasets}

