% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extractGT2NM}
\alias{extractGT2NM}
\title{Extract numeric data from genotype field of VCF}
\usage{
extractGT2NM(x)
}
\arguments{
\item{x}{A dataframe.}
}
\description{
Extract numeric data from genotype field of VCF
}

