% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{freq_peak}
\alias{freq_peak}
\title{freq_peak}
\usage{
freq_peak(myMat, pos, winsize = 10000L, bin_width = 0.02, count = FALSE)
}
\arguments{
\item{myMat}{a matrix of frequencies [0-1].}

\item{pos}{a numeric vector describing the position of variants in myMat.}

\item{winsize}{sliding window size.}

\item{bin_width}{Width of bins to summarize ferequencies in (0-1].}

\item{count}{logical specifying to count the number of non-NA values intead of reporting peak.}
}
\value{
A list
}
\description{
Find peaks in frequency data.
}
\details{
More to come.
}
\examples{
freqs <- matrix(runif(n=9), ncol=3, nrow=3)
pos <- 1:3
myPeaks <- freq_peak(freqs, pos)

data(vcfR_example)
ad <- extract.gt(vcf, element = "AD")
ad1 <- masplit(ad, record = 1)
ad2 <- masplit(ad, record = 2)
freqs <- ad1/(ad1+ad2)
myPeaks <- freq_peak(freqs, getPOS(vcf))
myPeaks <- freq_peak(freqs[1:115,], getPOS(vcf)[1:115])

# Visualize
mySample <- "P17777us22"
myWin <- 1
hist(freqs[myPeaks$wins[myWin,'START_row']:myPeaks$wins[myWin,'END_row'], mySample], 
     breaks=seq(0,1,by=0.02), col=8, main="", xlab="", xaxt="n")
axis(side=1, at=c(0,0.25,0.333,0.5,0.666,0.75,1), 
     labels=c(0,'1/4','1/3','1/2','2/3','3/4',1), las=3)
abline(v=myPeaks$peaks[myWin,mySample], col=2, lwd=2)


}

