% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_peak_plot.R
\name{freq_peak_plot}
\alias{freq_peak_plot}
\title{Plot freq_peak object}
\usage{
freq_peak_plot(pos, ab1 = NULL, ab2 = NULL, fp1 = NULL, fp2 = NULL,
  mySamp = 1, col1 = "#A6CEE344", col2 = "#1F78B444", mhist = TRUE, ...)
}
\arguments{
\item{pos}{chromosomal position of variants}

\item{ab1}{matrix of allele balances for allele 1}

\item{ab2}{matrix of allele balances for allele 2}

\item{fp1}{feq_peak object for allele 1}

\item{fp2}{feq_peak object for allele 2}

\item{mySamp}{sample indicator}

\item{col1}{color 1}

\item{col2}{color 2}

\item{mhist}{logical indicating to include a marginal histogram}

\item{...}{parameters passed on to other functions}
}
\value{
an invisible NULL.
}
\description{
Converts allele balance data produced by \code{freq_peak()} to a copy number by assinging the allele balance data (frequencies) to its closest expected ratio.
}
\details{
Converts allele balance data produced by \code{freq_peak()} to a copy number.
}
\examples{
data(vcfR_example)
freq_peak_plot(pos=1:40)

}
