% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_diff.R
\name{genetic_diff}
\alias{genetic_diff}
\title{Genetic differentiation}
\usage{
genetic_diff(vcf, pops, method = "nei")
}
\arguments{
\item{vcf}{a vcfR object}

\item{pops}{factor indicating populations}

\item{method}{the method to measure differentiation}
}
\description{
Calculate measures of genetic differentiation..
}
\details{
Measures of genetic differentiation, or fixation indicies, are commonly reported population genetic parameters.
This function reports genetic differentiation for all variants presented to it.


The method \strong{nei} returns Nei's Gst as well as Hedrick's correction for high alleleism (Hedrick 2005).
This method is based on heterozygosity where all of the alleles in a population are used to calculate allele frequecies.
This may make this a good choice when there is a mixture of ploidies in the sample.
}
\examples{
data(vcfR_example)
myPops <- as.factor(rep(c('a','b'), each = 9))
myDiff <- genetic_diff(vcf, myPops, method = "nei")
colMeans(myDiff[,c(3:6,9)], na.rm = TRUE)


}
\references{
Hedrick, Philip W. "A standardized genetic differentiation measure." Evolution 59.8 (2005): 1633-1638.
}
\seealso{
poppr.amova in \href{https://cran.r-project.org/package=poppr}{poppr}, amova in \href{https://cran.r-project.org/package=ade4}{ade4}, amova in \href{https://cran.r-project.org/package=pegas}{pegas}, \href{https://cran.r-project.org/package=hierfstat}{hierfstat}, \href{https://cran.r-project.org/package=poppr}{hierfstat}, \href{https://cran.r-project.org/package=DEMEtics}{DEMEtics}, and , \href{https://cran.r-project.org/package=mmod}{mmod}.
}
