% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NM2winNM}
\alias{NM2winNM}
\title{Extract numeric data from genotype field of VCF}
\usage{
NM2winNM(x, pos, maxbp, winsize = 100L)
}
\arguments{
\item{x}{A NumericMatrix}

\item{pos}{A vector of chromosomal positions}

\item{maxbp}{Length of chromosome}

\item{winsize}{Size (in bp) for windows}
}
\description{
Extract numeric data from genotype field of VCF
}

