% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordisample.R
\name{ordisample}
\alias{ordisample}
\title{Ordinate a sample's data}
\usage{
ordisample(x, sample, distance = "bray", plot = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{x}{an object of class vcfR or chromR.}

\item{sample}{a sample number where the first sample (column) is 2}

\item{distance}{metric to be used for ordination}

\item{plot}{logical specifying whether to plot the ordination}

\item{verbose}{logical specifying whether to produce verbose output}

\item{...}{parameters to be passed to child processes}
}
\value{
A list consisting of
}
\description{
Ordinate all available information from a sample's GT region and INFO column.
}
\details{
The INFO column of VCF data contains descriptors for each variant.
}
\examples{
\dontrun{
data(vcfR_test)
ordisample(vcfR_test, sample = 2)

myOrd <- ordisample(vcfR_test, sample = 2)
plot(myOrd$metaMDS, type = "n")
points(myOrd$metaMDS, display = "sites", pch=20, col="#8B4513")
text(myOrd$metaMDS, display = "spec", col="blue", ...)
plot(myOrd$envfit, col = "#008000", add = TRUE, ... )
}


}
\seealso{
\code{\link[vegan]{metaMDS}}
\code{\link[vegan]{vegdist}}
}

