% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_to_popsum.R
\name{gt2popsum}
\alias{gt2popsum}
\alias{Population genetics summaries}
\alias{gt.to.popsum}
\title{Population genetics summaries}
\usage{
gt2popsum(x, deprecated = TRUE)

gt.to.popsum(x)
}
\arguments{
\item{x}{object of class chromR or vcfR}

\item{deprecated}{logical specifying whether to run the function (FALSE) or present deprecation message (TRUE).}
}
\description{
Functions that make population genetics summaries
}
