\name{Ranking}
\alias{Ranking}
\alias{rank_variants}
\title{Ranking variants within windows}
\usage{
rank_variants_chrom(x, scores)
}
\arguments{
  \item{x}{an object of class Crhom or a data.frame
  containing...}

  \item{scores}{a vector of scores for each variant to be
  used to rank the data}
}
\description{
Rank variants within windows.
}

