% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readVcfHeader}
\alias{readVcfBody}
\alias{readVcfHeader}
\title{Read vcf header}
\usage{
readVcfHeader(x)

readVcfBody(x)
}
\arguments{
\item{x}{A String containing the filename}

\item{x}{A String containing the filename}
}
\description{
Read vcf header

Read vcf body
}

