% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcf_conversion.R
\name{Format conversion}
\alias{Format conversion}
\alias{vcfR2genind}
\alias{vcfR2loci}
\title{Convert vcf data to other formats}
\usage{
vcfR2genind(x, sep = "[|/]")

vcfR2loci(x)
}
\arguments{
\item{x}{an object of class Chrom or vcfR}

\item{sep}{character (to be used in a regular expression) to delimit the alleles of genotypes}
}
\description{
Convert vcfR object to other formats
}

