\docType{class}
\name{vcfR-class}
\alias{vcfR-class}
\title{vcfR class}
\description{
A class for storing vcf data.
}
\details{
Defines a class for variant call format data. A vcfR object
contains three slots.  The first slot is a character vector
which holds the meta data.  The second slot holds an eight
column data.frame to hold the fixed data.  The third slot
is a data.frame which holds the genotype data.
}

