\name{read.vcf}
\alias{read.vcf}
\title{read.vcf}
\usage{
read.vcf(x)
}
\arguments{
  \item{x}{variant call format (vcf) file}
}
\description{
Reads in a vcf file and stores it in a vcf class.
}
\details{
Reads in a vcf file and stores it in a vcf class.  Once the
number of lines the meta information contains the data is
divided into three tables: meta data, fixed data and
genotype data.
}

