% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfR2hapmap.R
\name{vcfR2hapmap}
\alias{vcfR2hapmap}
\title{Convert a vcfR object to hapmap}
\usage{
vcfR2hapmap(vcf)
}
\arguments{
\item{vcf}{a vcfR object.}
}
\value{
a data.frame that can be used as an input for GAPIT.
}
\description{
The function converts a vcfR object to hapmap
}
\details{
This function converts a vcfR object to a hapmat format.
}
\author{
Brian J. Knaus
}
