% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfR_example.R
\docType{data}
\name{vcfR_example_extra}
\alias{vcfR_example_extra}
\alias{seq_multi}
\alias{tab_multi}
\alias{vcf_multi}
\title{Extra example data for vcfR.}
\format{
A DNAbin object, a data.frame and a vcfR object
}
\description{
Example Data with multiple chromosomes, human data VCFs sections pulled 
from 1000 genomes. see data-raw/additional_test_data.R for details of object
generation
}
\details{
\itemize{
  \item seq_multi DNAbin object
  \item tab_multi gff format data.frame
  \item vcf_multi vcfR object
}
}
\examples{
data(vcfR_example_extra)
}
\keyword{datasets}
