% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vcfR_read_write.R
\name{read_write_vcf}
\alias{read.vcf}
\alias{read_write_vcf}
\alias{write.vcf}
\title{Read and write vcf files}
\usage{
read.vcf(x)

write.vcf(xvcf, vfile, mask = logical(0), APPEND = FALSE)
}
\arguments{
\item{x}{variant call format (vcf) file}

\item{xvcf}{a vcfR object}

\item{vfile}{an output filename}

\item{mask}{logical vector indicating rows to use}

\item{APPEND}{logical indicating whether to append to existing vcf file or write a new file}
}
\description{
Reads in a vcf file and stores it in a vcf class.
}
\details{
Reads in a vcf file and stores it in a vcf class.  Once the number of lines the meta information contains the data is divided into three tables: meta data, fixed data and genotype data.
}
\examples{
library(vcfR)
data(vcfR_example)
head(pinf_vcf)
plot(pinf_vcf)
pinf_vcf[1:6,]
}

