% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep.R
\docType{data}
\name{vep}
\alias{vep}
\title{Example data from the Variant Effect Predictor (VEP).}
\format{
A vcfR object
}
\description{
Example data to use with unit tests.
}
\details{
\itemize{
  \item vep vcfR object
}


Output from the \href{https://uswest.ensembl.org/info/docs/tools/vep/index.html}{VEP} may include values with multiple equals signs.
This does not appear to conform with the VCF specification (at the time of writing this \href{http://samtools.github.io/hts-specs/}{VCF v4.3}).
But it appears fairly easy to accomodate.
This example data can be used to make unit tests to validate functionality.
}
\examples{
data(vep)
vcfR2tidy(vep, info_only = TRUE)$fix


}
\keyword{datasets}
