% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Chrom_class.r
\docType{class}
\name{Chrom-class}
\alias{Chrom-class}
\title{Chrom class}
\description{
A class for representing chromosomes (or contigs).
}
\details{
Defines a class for chromosomal or contig data.

This object has a numbr of slots.

\itemize{
  \item name name of the object
  \item seq object of class DNAbin (ape)
  \item len length of sequence (integer)
  \item vcf.meta vcf meta data
  \item vcf.fix vcf fixed data
  \item vcf.gt vcf genotype data
  \item ann annotation data in a gff-like data.frame

  \item var.info a data.frame containing information on variants
  \item win.info a data.frame containing information on windows
  \item seq.info a list containing information on the sequence

}

More descriptions can be put here.
}

