\name{vcf2chrom}
\alias{ann2chrom}
\alias{vcf2chrom}
\alias{windowize}
\title{Chrom methods}
\usage{
vcf2chrom(x, y, ...)

ann2chrom(x, y, ...)

windowize(x, win.size = 1000, max.win = 10000)
}
\arguments{
  \item{x}{an object of class chrom}

  \item{y}{an object}

  \item{...}{arguments}

  \item{win.size}{window size, in base pairs}

  \item{max.win}{maximum window size}
}
\description{
Methods to work with objects of the chrom class
}
\details{
Reads in a vcf file and stores it in a Chrom class.

Reads in a vcf file and stores it in a Chrom class.

Reads in a vcf file and stores it in a Chrom class.
}

