% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Chrom_functions.R
\name{masker}
\alias{Chrom}
\alias{functions}
\alias{masker}
\alias{variant_table}
\alias{window_table}
\title{Chrom_functions}
\usage{
masker(x, QUAL = 999, mindp = 0.25, maxdp = 0.75, minmq = 20,
  maxmq = 50, ...)

variant_table(x)

window_table(x)
}
\arguments{
\item{x}{object of class Chrom}

\item{QUAL}{variant quality}

\item{mindp}{minimum cumulative depth}

\item{maxdp}{maximum cumulative depth}

\item{minmq}{minimum mapping quality}

\item{maxmq}{maximum mapping quality}

\item{...}{arguments to be passed to children functions}
}
\description{
Chrom_functions
}

