% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Chrom_functions.R
\name{Chrom functions}
\alias{Chrom functions}
\alias{extract_indels}
\alias{masker}
\alias{variant_table}
\alias{window_table}
\title{Chrom_functions}
\usage{
masker(x, QUAL = 999, mindp = 0.25, maxdp = 0.75, minmq = 20,
  maxmq = 50, ...)

variant_table(x)

window_table(x)

extract_indels(x, return_indels = FALSE)
}
\arguments{
\item{x}{object of class Chrom}

\item{QUAL}{minimum variant quality}

\item{mindp}{minimum cumulative depth}

\item{maxdp}{maximum cumulative depth}

\item{minmq}{minimum mapping quality}

\item{maxmq}{maximum mapping quality}

\item{...}{arguments to be passed to methods}

\item{return_indels}{logical indicating whether to return indels or not}
}
\description{
Functions which act on Chrom objects
}
\details{
The function \strong{masker} creates a logical vector that determines which variants are masked.
By masking certain variants, instead of deleting them, it preserves the dimensions of the data structure until a change needs to be committed.
Variants can be masked based on the value of the QUAL column of the vcf object.
Experience seems to show that this value is either at its maximum (999) or a rather low value.
The maximum and minimum sequence depth can also be used (mindp and maxdp).
The default is to mask all variants with depths of less than the 0.25 quantile and greater than the 0.75 quantile (these are also known as the lower and upper quartile).
The minimum and maximum mapping qualities (minmq, maxmq) can also be used.


This vector is stored in the var.info$mask slot of a Chrom object.

The function \strong{variant_table} creates a data.frame containing information about variants.

The funciton \strong{window_table}
}

