% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Chrom_methods.R
\docType{methods}
\name{show,Chrom-method}
\alias{ann2chrom}
\alias{chrom-methods}
\alias{create.chrom}
\alias{head,Chrom-method}
\alias{head.chrom}
\alias{names,Chrom-method}
\alias{names.chrom}
\alias{plot,Chrom-method}
\alias{plot.chrom}
\alias{proc.chrom}
\alias{show,Chrom-method}
\alias{vcf2chrom}
\alias{windowize}
\title{Chrom methods}
\usage{
\S4method{show}{Chrom}(object)

\S4method{names}{Chrom}(x)

\S4method{head}{Chrom}(x)

\S4method{plot}{Chrom}(x, y, ...)

vcf2chrom(x, y, ...)

ann2chrom(x, y, ...)

create.chrom(name, seq, vcf = NULL, ann = NULL, verbose = TRUE)

proc.chrom(x, verbose = TRUE, ...)

windowize(x, win.size = 1000, max.win = 10000)
}
\arguments{
\item{x}{an object of class Chrom}

\item{y}{an object}

\item{...}{arguments}

\item{name}{a name for the object}

\item{seq}{a sequence as a DNAbin object}

\item{vcf}{a vcfR object}

\item{ann}{an annotation file (gff-like)}

\item{verbose}{logical stating whether to produce verbose output}

\item{win.size}{window size, in base pairs}

\item{max.win}{maximum window size}

\item{x}{an object of class chrom}
}
\description{
Chrom methods

Methods to work with objects of the chrom class
}
\details{
Reads in a vcf file and stores it in a Chrom class.

Reads in a vcf file and stores it in a Chrom class.

Creates and names a chrom object from a name, a sequence (a DNA bin object),
vcf data (a vcfR object) and annotation data (gff-like).

Reads in a vcf file and stores it in a Chrom class.
}
\examples{
library(vcfR)
data(vcfR_example)
pinf_mt <- create.chrom('pinf_mt', seq=pinf_dna, vcf=pinf_vcf, ann=pinf_gff)
head(pinf_mt)
pinf_mt
names(pinf_mt)
plot(pinf_mt)
pinf_mt <- masker(pinf_mt)
pinf_mt <- proc.chrom(pinf_mt, win.size=1000)
plot(pinf_mt)

chromoqc(pinf_mt)
chromoqc(pinf_mt, xlim=c(25e+03, 3e+04), dot.alpha=99)

set.seed(10)
x1 <- as.integer(runif(n=20, min=1, max=39000))
y1 <- runif(n=length(x1), min=1, max=100)
chromodot(pinf_mt, x1=x1, y1=y1)
chromodot(pinf_mt, x1=x1, y1=y1, label1='My data', x2=x1, y2=y1, label2='More of my data', dot.alpha='ff')

chromohwe(pinf_mt, dot.alpha='ff')

chromopop(pinf_mt)
gt <- extract.gt(pinf_mt)
head(gt)
tab <- variant_table(pinf_mt)
head(tab)
win <- window_table(pinf_mt)
head(win)
}

