% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromR-method.R
\docType{methods}
\name{show,chromR-method}
\alias{chromR-method}
\alias{head,chromR-method}
\alias{names<-,chromR,character-method}
\alias{plot,chromR-method}
\alias{print,chromR-method}
\alias{show,chromR-method}
\title{chromR-method}
\usage{
\S4method{show}{chromR}(object)

\S4method{plot}{chromR}(x, y, ...)

\S4method{print}{chromR}(x, y, ...)

\S4method{head}{chromR}(x)

\S4method{names}{chromR,character}(x) <- value
}
\arguments{
\item{object}{an object of class chromR}

\item{x}{an object of class chromR}

\item{y}{some sort of object???}

\item{...}{Arguments to be passed to methods}

\item{value}{a character containing a name}
}
\description{
Methods that act on objects of class chromR
}
\details{
Methods that act on objects of class chromR.
}

