% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_chrom.R
\name{Create Chrom object}
\alias{Create Chrom object}
\alias{ann2chrom}
\alias{chrom-methods}
\alias{create_chrom}
\alias{seq2chrom}
\alias{vcf2chrom}
\title{Create Chrom object}
\usage{
create_chrom(title = "CHROM1", vcf, seq = NULL, ann = NULL,
  verbose = TRUE)

vcf2chrom(x, vcf)

seq2chrom(x, seq = NULL)

ann2chrom(x, gff)
}
\arguments{
\item{title}{a name for the chromosome (for plotting purposes)}

\item{vcf}{an object of class vcfR}

\item{seq}{a sequence as a DNAbin object}

\item{ann}{an annotation file (gff-like)}

\item{verbose}{should verbose output be printed to the console?}

\item{x}{an object of class Chrom}

\item{gff}{a data.frame containing annotation data in the gff format}
}
\description{
Creates and populates an object of class Chrom.
}
\details{
Creates and names a chrom object from a name, a chromosome (an ape::DNAbin object), variant data (a vcfR object) and annotation data (gff-like).
The function \strong{create_chrom} is a wrapper which calls functions to populate the slots of the Chrom object.

The function \strong{vcf2chrom} is called by create_chrome and transfers the data from the slots of a vcfR object to the slots of a Chrom object.
It also tries to extract the 'DP' and 'MQ' fileds (when present) from teh fix region's INFO column.
It is not anticipated that a user would need to use this function directly, but its placed here in case they do.

The function \strong{seq2chrom} is currently defined as a generic function.
This may change in the future.
This function takes an object of class DNAbin and assigns it to the 'seq' slot of a Chrom object.

The function \strong{ann2chrom} is called by create_chrome and transfers the information from a gff-like object to the 'ann' slot of a Chrom object.
It is not anticipated that a user would need to use this function directly, but its placed here in case they do.
}
\examples{
library(vcfR)
data(vcfR_example)
pinf_mt <- create_chrom('pinf_mt', seq=pinf_dna, vcf=pinf_vcf, ann=pinf_gff)
head(pinf_mt)
pinf_mt
names(pinf_mt)
plot(pinf_mt)
pinf_mt <- masker(pinf_mt)
pinf_mt <- proc_chrom(pinf_mt, win.size=1000)

plot(pinf_mt)

chromoqc(pinf_mt)
chromoqc(pinf_mt, xlim=c(25e+03, 3e+04), dot.alpha=99)

set.seed(10)
x1 <- as.integer(runif(n=20, min=1, max=39000))
y1 <- runif(n=length(x1), min=1, max=100)
chromodot(pinf_mt, x1=x1, y1=y1)

chromodot(pinf_mt, x1=x1, y1=y1, label1='My data',
          x2=x1, y2=y1, label2='More of my data',
          dot.alpha='ff')

chromohwe(pinf_mt, dot.alpha='ff')

chromopop(pinf_mt)
gt <- extract.gt(pinf_mt)
head(gt)
tab <- variant_table(pinf_mt)
head(tab)
win <- window_table(pinf_mt)
head(win)
}
\seealso{
\code{\link{Chrom-class}},
\code{\link{vcfR-class}},
\code{\link[ape]{DNAbin}},
\href{http://www.1000genomes.org/wiki/analysis/variant\%20call\%20format/vcf-variant-call-format-version-41}{vcf format},
\href{http://www.sequenceontology.org/gff3.shtml}{gff3 format}
}

