\docType{data}
\name{example_data}
\alias{example_data}
\alias{pinf_dna}
\alias{pinf_gff}
\alias{pinf_sc10}
\alias{pinf_sc100}
\alias{pinf_sc33}
\title{Example data for vcfR.}
\format{A data frame with 53940 rows and 10 variables}
\description{
An example dataset containing parts of the Phytophthora
infestans genome.
}
\details{
\itemize{ \item pinf_dna Super contigs 10, 33 and 100 as a
DNAbin object (ape) \item pinf_sc10 variant information for
supercontig 10 as a vcfR object \item pinf_sc100 variant
information for supercontig 100 as a vcfR object \item
pinf_sc33 variant information for supercontig 33 as a vcfR
object \item pinf_gff annotations in a gff format as a
data.frame

}
}
\keyword{datasets}

