% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract_gt.R
\name{extract.gt}
\alias{extract.gt}
\alias{extract_haps}
\alias{extract_indels}
\alias{extract_info}
\alias{get.alleles}
\alias{is_biallelic}
\alias{is_polymorphic}
\title{Extract elements from the GT section of a vcf format object}
\usage{
extract.gt(x, element = "GT", mask = FALSE, as.numeric = FALSE,
  return.alleles = FALSE, allele.sep = "/")

extract_indels(x, return_indels = FALSE)

extract_info(x, element, as.numeric = FALSE, mask = FALSE)

extract_haps(x, mask = FALSE, gt_split = "|", verbose = TRUE)

is_polymorphic(x, na.omit = FALSE)

is_biallelic(x)

get.alleles(x, split = "/", na.rm = FALSE, as.numeric = FALSE)
}
\arguments{
\item{x}{An object of class Chrom or vcfR}

\item{element}{element to extract from vcf genotype data. Common options include "DP", "GT" and "GQ"}

\item{mask}{a logical indicating whether to apply the mask (TRUE) or return all variants (FALSE). Alternatively, a vector of logicals may be provided.}

\item{as.numeric}{Logical, should the matrix be converted to numerics}

\item{return.alleles}{logical indicating whether to return the genotypes (0/1) or alleles (A/T)}

\item{allele.sep}{character which delimits the alleles in a genotype (/ or |)}

\item{return_indels}{logical indicating whether to return indels or not}

\item{gt_split}{character which delimits alleles in genotypes}

\item{verbose}{should verbose output be generated}

\item{na.omit}{logical to omit missing data}

\item{split}{character passed to strsplit to split the genotype into alleles}

\item{na.rm}{logical indicating whether to remove NAs}
}
\description{
Extract elements from the GT section of a vcf format object
}
\details{
Note that when 'as.numeric' is set to 'TRUE' but the data are not actually numeric, unexpected results will likely occur.
}

