% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract_gt.R
\name{extract.gt}
\alias{alleles_to_consensus}
\alias{extract.gt}
\alias{extract_haps}
\alias{extract_indels}
\alias{extract_info}
\alias{get.alleles}
\alias{is_biallelic}
\alias{is_polymorphic}
\title{Extract elements from the GT section of a vcf format object}
\usage{
extract.gt(x, element = "GT", mask = FALSE, as.numeric = FALSE,
  return.alleles = FALSE, allele.sep = "/")

extract_indels(x, return_indels = FALSE)

extract_info(x, element, as.numeric = FALSE, mask = FALSE)

extract_haps(x, mask = FALSE, gt_split = "|", verbose = TRUE)

is_polymorphic(x, na.omit = FALSE)

is_biallelic(x)

get.alleles(x, split = "/", na.rm = FALSE, as.numeric = FALSE)

alleles_to_consensus(x2, sep = "/", NA_to_n = TRUE)
}
\arguments{
\item{x}{An object of class Chrom or vcfR}

\item{element}{element to extract from vcf genotype data. Common options include "DP", "GT" and "GQ"}

\item{mask}{a logical indicating whether to apply the mask (TRUE) or return all variants (FALSE). Alternatively, a vector of logicals may be provided.}

\item{as.numeric}{Logical, should the matrix be converted to numerics}

\item{return.alleles}{logical indicating whether to return the genotypes (0/1) or alleles (A/T)}

\item{allele.sep}{character which delimits the alleles in a genotype (/ or |)}

\item{return_indels}{logical indicating whether to return indels or not}

\item{gt_split}{character which delimits alleles in genotypes}

\item{verbose}{should verbose output be generated}

\item{na.omit}{logical to omit missing data}

\item{split}{character passed to strsplit to split the genotype into alleles}

\item{na.rm}{logical indicating whether to remove NAs}

\item{x2}{a matrix of alleles as genotypes (e.g., A/A, C/G, etc.)}

\item{sep}{a character which delimits the alleles in a genotype (/ or |)}

\item{NA_to_n}{logical indicating whether NAs should be scores as n}
}
\description{
Extract elements from the GT section of a vcf format object
}
\details{
Note that when 'as.numeric' is set to 'TRUE' but the data are not actually numeric, unexpected results will likely occur.

The function \strong{extract.gt} isolates elements from the GT portion of vcf data.
Fields available for extraction are listed in the FORMAT column of the GT portion.
Because different vcf producing software produce different fields the options will vary by software.

The function \strong{extract_indels} is used to isolate indels from SNPs.
When the parameter return_indels is FALSE only SNPs will be returned.
When the parameter return_indels is TRUE only indels will be returned.

The function \strong{extract_info} is used to isolate elements from the INFO column of vcf data.

The function \strong{extract_haps} uses extract.gt to isolate genotypes.
It then uses the information in the REF and ALT columns as well as an allele delimiter (gt_split) to split genotypes into their allelic state.

The function \strong{is_polymorphic} returns a vector of logicals indicating whether a variant is polymorphic.
Variants in vcf files are always polymorphic.
However, if the variants are censored somehow (set to NA) or samples are removed a variant may become invariant.

The function \strong{is_bialleleic} returns a vector of logicals indicating whether a variant is biallelic.
Some analyses or downstream analyses only work with biallelic loci.
This function can help manage this.

The function \strong{get.alleles} takes a vector of genotypes and returns the unique alleles.

The function \strong{alleles_to_consensus} converts genotypes to a single consensus allele using IUPAC ambiguity codes for heterozygotes.
}

