\name{extractGT2NM}
\alias{extractGT2NM}
\title{Extract numeric data from genotype field of VCF}
\usage{
extractGT2NM(x, element = "DP")
}
\arguments{
  \item{x}{A dataframe}

  \item{element}{A string matching the element to be
  retrieved}
}
\description{
Extract numeric data from genotype field of VCF
}

