% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{freq_peak}
\alias{freq_peak}
\title{freq_peak}
\usage{
freq_peak(myMat, pos, winsize = 10000L, count = FALSE)
}
\arguments{
\item{myMat}{a matrix of frequencies [0-1].}

\item{pos}{a numeric vector describing the position of variants in myMat.}

\item{winsize}{sliding window size.}

\item{count}{logical specifying to count the number of non-NA values intead of reporting peak.}
}
\value{
A list
}
\description{
Find peaks in frequency data.
}
\details{
More to come.
}
\examples{
freqs <- matrix(runif(n=9), ncol=3, nrow=3)
pos <- 1:3
myPeaks <- freq_peak(freqs, pos)

data(vcfR_example)
ad <- extract.gt(vcf, element = "AD")
ad1 <- masplit(ad, record = 1)
ad2 <- masplit(ad, record = 2)
freqs <- ad1/(ad1+ad2)
myPeaks <- freq_peak(freqs, getPOS(vcf))


}

