% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{heatmap-barplots}
\alias{heatmap-barplots}
\alias{heatmap.bp}
\title{Heatmap with barplots}
\usage{
heatmap.bp(x, cbarplot = TRUE, rbarplot = TRUE, legend = TRUE,
  clabels = TRUE, rlabels = TRUE, na.rm = TRUE, scale = c("row",
  "column", "none"), col.ramp = colorRampPalette(c("red", "yellow",
  "#008000"))(100), ...)
}
\arguments{
\item{x}{a numeric matrix.}

\item{cbarplot}{a logical indicating whether the columns should be summarized with a barplot.}

\item{rbarplot}{a logical indicating whether the rows should be summarized with a barplot.}

\item{legend}{a logical indicating whether a legend should be plotted.}

\item{clabels}{a logical indicating whether column labels should be included.}

\item{rlabels}{a logical indicating whether row labels should be included.}

\item{na.rm}{a logical indicating whether missing values should be removed.}

\item{scale}{character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default is "none".}

\item{col.ramp}{vector of colors to be used for the color ramp.}

\item{...}{additional arguments to be passed on.}
}
\description{
Heatmap of a numeric matrix with barplots summarizing columns and rows.
}
\details{
The function heatmap.bp creates a heatmap from a numeric matrix with optional barplots to summarize the rows and columns.
}
\examples{
library(vcfR)
data(vcfR_example)
pinf_mt <- create_chrom('pinf_mt', seq=pinf_dna, vcf=pinf_vcf, ann=pinf_gff)
pinf_mt <- masker(pinf_mt)
pinf_gq <- extract.gt(pinf_mt, element="GQ", as.numeric=TRUE)
heatmap.bp(pinf_gq)
heatmap.bp(pinf_gq, scale="col")
heatmap.bp(pinf_gq, col.ramp = colorRampPalette(c("red", "yellow", "#008000"))(100))
heatmap.bp(pinf_gq, col.ramp = colorRampPalette(c("#D55E00", "#F0E442", "#009E73"))(100))
}
\seealso{
heatmap, image, heatmap2 in library(gplots).
}

