% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/io_vcfR.R
\name{read.vcf.R}
\alias{memory_plot}
\alias{read.vcf}
\alias{read.vcf.R}
\alias{write.vcf}
\alias{write.vcf.R}
\alias{write.vcf.gz}
\alias{write_var_info}
\alias{write_win_info}
\title{Read and write vcf format files}
\usage{
read.vcf.R(file)

write.vcf.R(x, file = "", mask = logical(0), APPEND = FALSE)

read.vcf(file, limit = 1e+09)

write.vcf(x, file = "", mask = FALSE, APPEND = FALSE)

write.vcf.gz(x, file = "", mask = FALSE, APPEND = FALSE)

memory_plot(exponent_range = 2:6)

write_var_info(x, file = "", mask = FALSE, APPEND = FALSE)

write_win_info(x, file = "", APPEND = FALSE)
}
\arguments{
\item{file}{A filename for a variant call format (vcf) file}

\item{x}{An object of class vcfR or Chrom}

\item{mask}{logical vector indicating rows to use}

\item{APPEND}{logical indicating whether to append to existing vcf file or write a new file}

\item{limit}{amount of memory not to exceed when reading in a file}

\item{exponent_range}{range of values to be used for object sizes (10^exponent_range)}
}
\description{
Read and write files in the vcf format.
}
\details{
Reads in a vcf file and stores it in a vcf class.  Once the number of lines the meta information contains the data is divided into three tables: meta data, fixed data and genotype data.

\code{read.vcf.devel} is a custom C++ implementation which may be faster than read.vcf.
}
\examples{
library(vcfR)
data(vcfR_example)
head(pinf_vcf)
plot(pinf_vcf)
pinf_vcf[1:6,]
}
\seealso{
Bioconductor: \href{www.bioconductor.org/packages/3.0/bioc/html/VariantAnnotation.html}{VariantAnnotation::readVcf}
}

