\name{NM2winNM}
\alias{NM2winNM}
\title{Extract windows of numeric data from genotype field of VCF}
\usage{
NM2winNM(x, pos, maxbp, winsize = 100L)
}
\arguments{
  \item{x}{A NumericMatrix}

  \item{pos}{A vector of chromosomal positions}

  \item{maxbp}{Length of chromosome}

  \item{winsize}{Size (in bp) for windows}
}
\description{
Extract windows of numeric data from genotype field of VCF
}

