% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordisample.R
\name{ordisample}
\alias{ordisample}
\title{Ordinate a sample's data}
\usage{
ordisample(x, sample, distance = "bray", plot = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{x}{an object of class vcfR or chromR.}

\item{sample}{a sample number where the first sample (column) is 2}

\item{distance}{metric to be used for ordination, options are in \code{\link[vegan]{vegdist}}}

\item{plot}{logical specifying whether to plot the ordination}

\item{verbose}{logical specifying whether to produce verbose output}

\item{...}{parameters to be passed to child processes}
}
\value{
A list consisting of two objects.
The first object is an object of class 'metaMDS' created by the function vegan::metaMDS.
The second is an object of class 'envfit' created by the function vegan::envfit.
}
\description{
Ordinate information from a sample's GT region and INFO column.
}
\details{
The INFO column of VCF data contains descriptors for each variant.
Each sample typically includes several descriptors of each variant in the GT region as well.
This can present an overwhelming amount of information.
Ordination is used in this function to reduce this complexity.

The ordination procedure can be rather time consuming depending on how much data is used.
I good recommendation  is to always start with a small subset of your full dataset and slowly scale up.
There are several steps in this function that attempt to eliminate variants or characters that have missing values in them.
This that while starting with a small number is good, you will need to have a large enough number so that a substantial amount of the data make it to the ordination step.
In the example I use 100 variants which appears to be a reasonable compromise.


Several of the steps of this ordination make distributional assumptions.
That is, they assume the data to be normally distributed.
There is no real reason to assume this assumption to be valid with VCF data.
It has been my experience that this assumption is frequently violated with VCF data.
It is therefore suggested to use this funciton as an exploratory tool that may help inform other decisions.
These analyst may be able to address these issues through data transformation or other topics beyond the scope of this function.
Interpretation of the results of this function need to take into account that assumptions may hav been violated.
}
\examples{
\dontrun{
data(vcfR_example)
ordisample(vcf[1:100,], sample = 2)

myOrd <- ordisample(vcf[1:100,], sample = 2, plot = FALSE)
names(myOrd)
plot(myOrd$metaMDS, type = "n")
points(myOrd$metaMDS, display = "sites", pch=20, col="#8B4513")
text(myOrd$metaMDS, display = "spec", col="blue")
plot(myOrd$envfit, col = "#008000", add = TRUE)
}


}
\seealso{
\code{\link[vegan]{metaMDS}}
\code{\link[vegan]{vegdist}}
}

