% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ranking.R
\name{Ranking}
\alias{Ranking}
\alias{rank_variants_chromR}
\title{Ranking variants within windows}
\usage{
rank_variants_chromR(x, scores)
}
\arguments{
\item{x}{an object of class Crhom or a data.frame containing...}

\item{scores}{a vector of scores for each variant to be used to rank the data}
}
\description{
Rank variants within windows.
}

