% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{Ranking}
\alias{Ranking}
\alias{rank.variants.chromR}
\title{Ranking variants within windows}
\usage{
rank.variants.chromR(x, scores)
}
\arguments{
\item{x}{an object of class Crhom or a data.frame containing...}

\item{scores}{a vector of scores for each variant to be used to rank the data}
}
\description{
Rank variants within windows.
}

