% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tables.R
\name{write.var.info}
\alias{write.var.info}
\alias{write.win.info}
\title{Write summary tables from chromR objects}
\usage{
write.var.info(x, file = "", mask = FALSE, APPEND = FALSE)

write.win.info(x, file = "", APPEND = FALSE)
}
\arguments{
\item{x}{An object of class chromR}

\item{file}{A filename for the output file}

\item{mask}{logical vector indicating rows to use}

\item{APPEND}{logical indicating whether to append to existing file (omitting the header) or write a new file}
}
\description{
Write summary tables from chromR objects.
}
\details{
The function \strong{write.var.info} takes the variant information table from a chromR object and writes it as a comma delimited file. 

The function \strong{write.win.info} takes the window information table from a chromR object and writes it as a comma delimited file.
}
\seealso{
\code{\link{write.vcf}}
}

