% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcf_conversion.R
\name{Format conversion}
\alias{Format conversion}
\alias{vcfR2DNAbin}
\alias{vcfR2genind}
\alias{vcfR2loci}
\title{Convert vcfR objects to other formats}
\usage{
vcfR2genind(x, sep = "[|/]")

vcfR2loci(x)

vcfR2DNAbin(x, extract.indels = TRUE, consensus = TRUE,
  extract.haps = FALSE, gt.split = "|", ref.seq = NULL,
  start.pos = NULL, verbose = TRUE)
}
\arguments{
\item{x}{an object of class Chrom or vcfR}

\item{sep}{character (to be used in a regular expression) to delimit the alleles of genotypes}

\item{extract.indels}{logical, at present, the only option is TRUE}

\item{consensus}{logical, at present, the only option is TRUE}

\item{extract.haps}{logical specifying whether to separate genotype into alleles based on a delimiting character}

\item{gt.split}{character to delimit alleles within genotypes}

\item{ref.seq}{reference sequence for the region being converted}

\item{start.pos}{chromosomal position for the start of the ref.seq}

\item{verbose}{logical specifying whether to produce verbose output}
}
\description{
Convert vcfR objects to objects supported by other packages
}
\details{
After processing vcf data in vcfR, one will likely proceed to an analysis step.
Within R, three obvious choices are:
\href{http://cran.r-project.org/web/packages/pegas/index.html}{pegas},
\href{http://cran.r-project.org/web/packages/adegenet/index.html}{adegenet}
and \href{http://cran.r-project.org/web/packages/poppr/index.html}{poppr}.
The package pegas uses objects of type loci.
The function vcfR2loci calls extract.gt to create a matrix of genotypes which is then converted into an object of type loci.
The packages adegenet and poppr use the genind object.
The function vcfR2genind uses extract.gt to create a matrix of genotypes and uses the adegenet function df2genind to create a genind object.
The package poppr additionally uses objects of class genclone.
A genind object can be converted to a genclone object with the function poppr::as.genclone.

The DNAbin object stores nucleotide sequence information.
This means that in order to convert vcf data to a nucleotide representation.
For haploid data, this is straight forward.
For diploid data there is the option of converting heterozygous sites to IUPAC ambiguity codes.
This results in one sequence per diploid individual.
Note that functions called downstream of this choice may handle IUPAC ambiguity codes in unexpected manners.
If you have phased data, an alternative is to extract the haplotypes from each genotype.
This should work for diploids and higher ploids.

A multiple sequence alignment for an entire chromosome contains a high degree of redundancy in invariant sites.
This is why vcf files contain only information on variants.
However, many analyses make use of invariant sites.
For example, likelihood inference of phylogeny (and therefore Bayesian methods as well) requires invariant sites.
By default, the function \strong{vcfR2DNAbin} by default will return an object of class DNAbin which only contains the variable positions.
In order to fill in the invariant sites, the parameter ref.seq should be provided.
When vcfR2DNAbin encounters missing data in the vcf data (NA) it is coded as an ambiguous nucleotide (n) in the DNAbin object.
Providing an entire chromosome may exceed available memory.
By providing a ref.seq for a region (of class ape::DNAbin), for example a gene, and the start.pos for that sequence (the end position will be determined from the sequence length), complete sequences can be created.
}
\seealso{
\code{\link{extract.gt}},
\code{\link[adegenet]{df2genind}},
\code{\link[adegenet]{genind}},
\href{http://cran.r-project.org/web/packages/pegas/index.html}{pegas},
\href{http://cran.r-project.org/web/packages/adegenet/index.html}{adegenet},
and
\href{http://cran.r-project.org/web/packages/poppr/index.html}{poppr}.
}

