% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcf_conversion.R
\name{Format conversion}
\alias{Format conversion}
\alias{gt_to_df}
\title{Convert vcf data to other formats}
\usage{
gt_to_df(x)
}
\arguments{
\item{x}{A matrix of genotypes}
}
\description{
Convert vcfR object to other formats
}

