% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vcf.R
\docType{class}
\name{vcfR-class}
\alias{vcfR-class}
\title{vcfR class}
\description{
A class for storing vcf data.
}
\details{
Defines a class for variant call format data.
A vcfR object contains three slots.  The first slot
is a character vector which holds the meta data.  The
second slot holds an eight column data.frame to hold the
fixed data.  The third slot is a data.frame which holds
the genotype data.
}
\section{Slots}{

\describe{
\item{\code{meta}}{character vector for the meta (header) information}

\item{\code{fix}}{data.frame for the fixed information}

\item{\code{gt}}{data.frame for the genotype information}
}}

