% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{vcfR-class}
\alias{vcfR-class}
\title{vcfR class}
\description{
An S4 class for storing VCF data.
}
\details{
Defines a class for variant call format data.
A vcfR object contains three slots.
The first slot is a character vector which holds the meta data.
The second slot holds an eight column matrix to hold the fixed data.
The third slot is a matrix which holds the genotype data.
The genotype data is optional according to the VCF definition.
When it is missing the gt slot should consist of a character matrix with zero rows and columns.


See \code{vignette('vcf_data')} for more information.
See the \href{http://samtools.github.io/hts-specs/}{VCF specification} for the file specification.
}
\section{Slots}{

\describe{
\item{\code{meta}}{character vector for the meta information}

\item{\code{fix}}{matrix for the fixed information}

\item{\code{gt}}{matrix for the genotype information}
}}

