% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vcf.R
\docType{methods}
\name{head,vcfR-method}
\alias{[,vcfR-method}
\alias{[]}
\alias{head,vcfR-method}
\alias{head.vcfR}
\alias{plot,vcfR-method}
\alias{plot.vcfR}
\alias{read.vcf}
\alias{subset,vcfR-method}
\alias{subset.vcf}
\alias{write.vcf}
\title{vcfR methods}
\usage{
\S4method{head}{vcfR}(x, n = 6, ...)

\S4method{plot}{vcfR}(x, y, ...)

\S4method{subset}{vcfR}(x, regex, ...)

\S4method{[}{vcfR}(x, i, j, drop)

read.vcf(x)

write.vcf(xvcf, vfile, mask = logical(0), APPEND = FALSE)
}
\arguments{
\item{x}{variant call format (vcf) file}

\item{n}{number of rows to print}

\item{...}{Arguments to be passed to methods}

\item{y}{not used}

\item{regex}{a regular expression to search for}

\item{i}{vector of rows (variants) to include}

\item{j}{vector of columns (samples) to include}

\item{drop}{delete the dimensions of an array which only has one level}

\item{xvcf}{a vcfR object}

\item{vfile}{an output filename}

\item{mask}{logical vector indicating rows to use}

\item{APPEND}{logical indicating whether to append to existing vcf file or write a new file}
}
\description{
Reads in a vcf file and stores it in a vcf class.
}
\details{
Reads in a vcf file and stores it in a vcf class.  Once the number of lines the meta information contains the data is divided into three tables: meta data, fixed data and genotype data.
}
\examples{
library(vcfR)
data(vcfR_example)
head(pinf_vcf)
plot(pinf_vcf)
pinf_vcf[1:6,]
}

