\docType{package}
\name{vcfR}
\alias{vcfR}
\alias{vcfR-package}
\title{Variant call format files processed with vcfR.}
\description{
vcfR
}
\details{
Lubridate provides tools that make it easier to parse and
manipulate dates. These tools are grouped below by common
purpose. More information about each function can be found
in its help documentation.
}
\references{
Brian Knaus (2014). Variant call format files processed
with vcfR. Journal of Statistical Software, 40(3), 1-25.
\url{http://www.jstatsoft.org/v40/i03/}.
}

