\docType{package}
\name{vcfR}
\alias{vcfR}
\alias{vcfR-package}
\title{Variant call format files processed with vcfR.}
\description{
vcfR
}
\details{
The complete list of functions can be displayed with
"library(help = vcfR)".

Lubridate provides tools that make it easier to parse and
manipulate dates. These tools are grouped below by common
purpose. More information about each function can be found
in its help documentation.
}
\references{
Brian Knaus (2014). Variant call format files processed
with vcfR. Journal TBA, NN(N), N-NN.
\url{http://www.jstatsoft.org/v40/i03/}.
}

