% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfR2hapmap.R
\name{vcfR2hapmap}
\alias{vcfR2hapmap}
\title{Convert a vcfR object to hapmap}
\usage{
vcfR2hapmap(vcf)
}
\arguments{
\item{vcf}{a vcfR object.}
}
\value{
a text file that can be used as an input for MigrateN software (SNP format).
}
\description{
The function converts a vcfR object to hapmap
}
\details{
This function converts a vcfR object to a text file which can be used as input for MigrateN.
}
\seealso{
\href{http://popgen.sc.fsu.edu/Migrate/Migrate-n.html}{Migrate-N} website.
}
\author{
Brian J. Knaus
}
