% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfR_example.R
\docType{data}
\name{vcfR_example}
\alias{vcfR_example}
\alias{dna}
\alias{gff}
\alias{vcf}
\title{Example data for vcfR.}
\format{A DNAbin object, a data.frame and a vcfR object}
\description{
An example dataset containing parts of the *Phytophthora infestans* genome.
}
\details{
\itemize{
  \item dna DNAbin object
  \item gff gff format data.frame
  \item vcf vcfR object
}

This data is a subset of the pinfsc50 dataset.
It has been subset to positions between 500 and 600 kbp.
The coordinate systems of the vcf and gff file have been altered by subtracting 500,000.
This results in a 100 kbp section of supercontig_1.50 that has positional data ranging from 1 to 100 kbp.

Note that it is encouraged to keep package contents small to facilitate easy
downloading and installation.  This is why a mitochondrion was chosen as an
example.  In practice I've used this package on supercontigs.  This package
was designed for much larger datasets in mind than in this example.
}
\examples{
data(vcfR_example)




}
\keyword{datasets}
