% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfR_test.R
\docType{data}
\name{vcfR_test}
\alias{vcfR_test}
\alias{vcf_test}
\title{Test data for vcfR.}
\format{A vcfR object}
\description{
A test file containing a diversity of examples intended to test functionality.
}
\details{
\itemize{
  \item vcfR_test vcfR object
}


This data set began as the example (section 1.1) from The Variant Call Format Specification \href{http://samtools.github.io/hts-specs/}{VCFv4.3} .
This data consisted of 3 samples and 5 variants.
As I encounter examples that challenge the code in vcfR they can be added to this data set.
}
\examples{
data(vcfR_test)


\dontrun{
  # When I add data it can be saved with this command.
  save(vcfR_test, file="data/vcfR_test.RData") 
  }


}
\keyword{datasets}
